bench: moved statistics fn to print_utils
